pyproject.toml
[tool.poetry]
name = "ai-chat-assistant"
version = "1.0.0"
description = "AI Chat Assistant with multiple models support"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flet = "^0.7.0"
requests = "^2.28.0"
python-dotenv = "^0.21.0"
pandas = "^1.5.0"
plotly = "^5.13.0"
psutil = "^5.9.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
black = "^22.3.0"
flake8 = "^4.0.1"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
